Fix bug in Plugin class's clean method to check if .dirty file exists before removing it

+++ src/compose_web_manager/plugin.py
@@ -44,7 +44,8 @@ class Plugin:
       """
       Mark a plugin as clean.
       """
-      os.remove(f'{SB_COMPOSE_ROOT}/{self.name}/.dirty')
+      if os.path.exists(f'{SB_COMPOSE_ROOT}/{self.name}/.dirty'):
+        os.remove(f'{SB_COMPOSE_ROOT}/{self.name}/.dirty')
       
   def restart(self):
       """
